requirements.txt
numpy==1.14.5
tensorflow==1.10.0
scikit-image==0.13.1
scikit-learn==0.19.1
Pillow==8.0.1
tqdm==4.46.1
six==1.15.0
opencv-python==4.2.0.34
